Pequenos incrementos e ajustes

+++ listareliminar/listardepartamentos.php
@@ -19,16 +19,21 @@ $consulta = 'SELECT * FROM departamento';
 if (!$resultado = $ligacao->query($consulta)) {
     echo ' Falha na consulta: '. $ligacao->error;
     $ligacao->close();  /* fechar a ligação */
+    exit();
 }
-else{
+    
+if($ligacao->affected_rows == 0){
+   echo " Não existem registos na tabela departamento ";
+}
+else{  
     /* percorrer os registos (linhas) da tabela e mostrar na página */
     while ($row = $resultado->fetch_assoc()){   
         echo 'id: ' . $row['id'] . ' Nome: ' . $row['nome'] . 
              '<a href="eliminardepartamento.php?id='.$row['id'] .
              '&operacao=eliminar"> - Eliminar</a><br>';
     }
-
+}
 $resultado->free();      /* libertar o resultado */
 $ligacao->close();       /* fechar a ligação */
-}
+
